restruct and func for edges added
