+++ entity/model.hpp
@@ -1,0 +1,9 @@
+#ifndef MODEL_H
+#define MODEL_H
+
+#include "entity.h"
+#include <string>
+#include <vector>
+using std::vector;
+
+#endif // MODEL_H
